AST: Split off ModuleNameLookup.h from NameLookup.h
